Done - implemented top3-based-2 algorithm - edited the type definition of variable probability to float TODO - test the algorithm by using evaluation metrics

+++ registration/registration_system/grade_priority_system.py
@@ -16,7 +16,7 @@ class GradePrioritySystem(LotterySystem):
         """
         super().__init__(courses_dict)
 
-    def register_students(self, students: List[Student], graduate_standard: int, probability: int) -> List[Student]:
+    def register_students(self, students: List[Student], graduate_standard: int, probability: float) -> List[Student]:
         """Register students to courses
 
         Args:
